requirements.txt
asgiref==3.2.7
astroid==2.5.1
backports.entry-points-selectable==1.1.0
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
colorama==0.4.4
cycler==0.10.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.0.5
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-cors-headers==3.10.0
django-crispy-forms==1.11.1
django-environ==0.4.5
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.7.0
joblib==1.0.1
Keras==2.4.3
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.0
platformdirs==2.2.0
psycopg2-binary==2.8.6
pylint==2.7.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
six==1.16.0
sqlparse==0.3.1
sweetify==2.0.0
threadpoolctl==2.2.0
toml==0.10.2
urllib3==1.26.4
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
wrapt==1.12.1
